I added int() in number 2, which ran calculations and returned an output instead of printing a string
In part three I fixed:
Needed int before input() in line 10,
Lack of parentheses () after print in line 13,
The variables "number1" and "number2" were not the same as the variables "Number1" or "Number2",
Quotations after Print( before "Seven", after "is"),
and 7(number2) was invalid syntax.

I made questions and variables for the base and height of the triangle, using int(input()) = variable, and then plugged the variables into a formula for the area of a triangle.

I had trouble at the end with the commas, but came around to them through trial and error. I think the first part of the assignment helped, because thinking about + and *, as well as how there were commas after quotations led to the conclusion that commas were necessary and created spaces between variables representing strings. Also important is that the variables for names did not need int(), because the variables are supposed to record strings.